% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{mungebit-class}
\alias{mungebit}
\alias{mungebit-class}
\title{This class is intended to abstract the idea of separate data preparation
during training versus prediction.}
\description{
This class is intended to abstract the idea of separate data preparation
during training versus prediction.
}

