% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mungebit__predict}
\alias{mungebit__predict}
\title{Run the predict function on a mungebit.}
\usage{
mungebit__predict(mungeplane, ...)
}
\arguments{
\item{mungeplane}{mungeplane. Essentially an environment containing
a \code{data} variable.}

\item{...}{additional arguments to the mungebit's \code{predict_fn}.}
}
\description{
Run the predict function on a mungebit.
}
\seealso{
\code{\link{mungebit__run}}, \code{\link{mungebit__initialize}}
}

