% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mungebit__run}
\alias{mungebit__run}
\title{Run a mungebit.}
\usage{
mungebit__run(mungeplane, ...)
}
\arguments{
\item{mungeplane}{mungeplane. Essentially an environment containing
a \code{data} variable.}

\item{...}{additional arguments to the mungebit's \code{train_fn} or
\code{predict_fn}.}
}
\description{
Imagine flipping a switch on a set of train tracks. A mungebit
behaves like this: once the \code{trained} switch is flipped,
it can only run the \code{predict_fn}, otherwise it will
run the \code{train_fn}.
}
\seealso{
\code{\link{mungebit__initialize}}
}

