% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/mungebits-package.r
\docType{package}
\name{mungebits}
\alias{mungebits}
\alias{mungebits-package}
\title{Mungebits is a collection of data processing snippets for the
munger library.}
\description{
Mungebits is a collection of data processing snippets for the
munger library.
}

