% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungeplane.r
\name{mungeplane}
\alias{mungeplane}
\title{Mungeplane are reference classes wrappers around a dataframe.}
\usage{
mungeplane(dataframe)
}
\arguments{
\item{dataframe}{a \code{data.frame}. This can also be a list to help with
optimization procedures in mungebit operations that temporarily remove
the \code{data.frame} class to use built-in C replacement functions.}
}
\value{
mungeplane an environment with a "data" object of class "mungeplane"
}
\description{
By exploiting pass-by-reference semantics of reference classes,
mungeplanes allow mungebits to manipulate individual columns and rows of
dataframes without forcing copy-on-write triggers that would copy the
entire dataframe.
}
\examples{
\dontrun{
mp <- mungeplane(iris)
(function(plane) plane$data[[1]] <- 2 * plane$data[[1]])(mp)
stopifnot(all.equal(mp$data[[1]], 2 * iris[[1]]))
}
}

