% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungebits-package.r
\docType{package}
\name{mungebits}
\alias{mungebits}
\alias{mungebits-package}
\title{Mungebits is a collection of data processing snippets for the
munger library.}
\description{
Mungebits is a collection of data processing snippets for the
munger library.
}

